API 아티스트프로필조회 수정
